Reorganize vasu util code and fix published at label
